requirements.txt
h5py==2.10.0
numpy==1.18.5
tensorflow-gpu==2.2.0
opencv-python==4.5.1.48
Pillow==8.1.0
Flask==1.1.2
scikit-image==0.18.1
scikit-learn==0.24.0
selenium==3.141.0